Fixes for OSX Mavericks
Panda is in pretty bad shape since it depends on
carbon, which is now fully deprecated and removed
in OSX 10.9. This means that earlier versions of
Panda - pre-1.9 - don't behave at all for window
events. Building Panda3d with librocket support
on 64 bit Python on Mavericks is not working yet
so librocket elements have been removed.